To have the data to be watched there
I've wrote a docker compose that creates the container, download your file, and synchronize an external volume into the container to watch what's happening on that directory. The only thing I need to avoid changing this file during the container build, is to have it pointing to the right directory. I'm using jupyter/base-notebook container to build this.